pyproject.toml
[project]
name = "battery-symbols"
version = "0.1.0"
description = "Battery symbols and font"
authors = [
    {name = "Chris Waltrip",email = "chris@walt.rip"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "svg-py (>=1.6.0,<2.0.0)",
    "ufolib2 (>=0.17.1,<0.18.0)",
    "fonttools (>=4.58.0,<5.0.0)",
    "defcon (>=0.12.1,<0.13.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
commitizen = "^4.7.2"
setuptools = "^80.7.1"

